# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger 커피공장 플랫폼"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
...
schemes:
  - "https"
  - "http"
paths:
...
  /api/banner/cancelDeploy:
    put:
      tags:
        - "배너관리"
      summary: "배너 배포취소"
      description: ""
      operationId: "cancelDeploy"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "body"
          in: "body"
          description: "배너 배포취소를 BannerDTO로 받는다"
          required: true
          schema:
            $ref: "#/definitions/BannerDTO"  
      responses:
        "200":
          description: "successful operation"
        "401":
          description: "Invalid input"  
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  BannerDTO:
    type: "object"
    properties:
      bannerId:
        type: "string"
        format: "string"
      name:
        type: "string"
        format: "string"      
      imgUrl:
        type: "string"
        format: "string"      
      width:
        type: "number"
        format: "number"      
      height:
        type: "number"
        format: "number"      
      location:
        type: "string"
        format: "string"      
      startDate:
        type: "string"
        format: "string"      
      endDate:
        type: "string"
        format: "string"      
      title:
        type: "string"
        format: "string"      
      description:
        type: "string"
        format: "string"      
      sort:
        type: "number"
        format: "number"      
      deployYn:
        type: "string"
        format: "string"      
...